+++ source/rafcon/utils/ros2/__init__.py
@@ -1,0 +1,12 @@
+# Copyright (C) 2016-2026 DLR
+#
+# All rights reserved. This program and the accompanying materials are made
+# available under the terms of the Eclipse Public License v1.0 which
+# accompanies this distribution, and is available at
+# http://www.eclipse.org/legal/epl-v10.html
+#
+# Contributors:
+# Johannes Ernst
+# Ryo Sakagami
+#
+# NOTE: To use the scripts, ROS2 needs to be installed (tested with 'Humble')