+++ tests/assets/unit_test_state_machines/start_from_here_test/start_from_here_test/example_hierarchy_GFPACD/ExecutionState_4_JVJFXI/script.py
@@ -1,0 +1,4 @@
+
+def execute(self, inputs, outputs, gvm):
+
+    return "success"